Started split of 6502 opcodes into several files
The PC relative addressing came to it's limits when I implemented the 6502 opcodes in one assembly
file. Putting them into several files solves this problem. Additionally it's a bit easier to find a
certain opcode now. Additionally this patch contains the implementation for AND.

committer: Thomas Giesel <skoe@directbox.com>