config files added
The old, simple way specifying options on the compiler command line ("-D") is not satisfying when there are
lots of them. Basically the idea was taken from sd2iec, but as the awk script did not do exactly what I
wanted, I re-implemented it in sed.

committer: Thomas Giesel <skoe@directbox.com>